new version for orchestrator split compare
